pyproject.toml
[tool.poetry]
name = "ROILoc"
version = "0.2.0"
description = "A simple package to center and crop T1w & T2w MRIs around a given region of interest by its name."
license = "MIT"
readme = "README.rst"
authors = ["Clément POIRET <poiret.clement@outlook.fr>"]
keywords = ["mri", "brain", "t1w", "t2w", "registration"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Image Processing"
]
homepage = "https://hippomnesis.dev"
repository = "https://github.com/clementpoiret/ROILoc"

[tool.poetry.dependencies]
python = "^3.7"
antspyx = "^0.2.7"
pandas = "^1.3.0"
rich = "^10.5.0"
importlib-resources = "^5.2.0"

[tool.poetry.scripts]
roiloc = 'roiloc.roiloc:start'

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
